updated code

+++ tests/Linear Equations/test_block_two.py
@@ -8,14 +8,15 @@ A = tf.constant([
     [2, 1, 1],
     [0, 5, 1],
     [-1, 2, 1]
-], dtype='float64')
+], dtype=tf.float64)
 
-B = -tf.constant([[10], [30], [10]])
+B = -tf.constant([[10], [30], [10]], dtype=tf.float64)
 
 X = tf.matrix_solve(A, B)
 
 with tf.Session() as session:
     result = session.run(X)
     D, E, F = result.flatten()
+    # result should be Equation: x**2 + y**2 + -2.0x + -6.0y + 0.0 = 0
 
     print("Equation: x**2 + y**2 + {D}x + {E}y + {F} = 0".format(**locals()))
